fix(pi-sequential-thinking): move EffectiveConfigStatus to config.ts
Review feedback (P1, maintainability). config.ts was importing
EffectiveConfigStatus from storage.ts — the dependency arrow was
backwards. The type describes config-resolution output (storageDir,
maxBytes, maxLines, sources), so it belongs in config.ts. storage.ts
now imports it (for getStatus's optional argument shape) instead of
defining it.

Also apply two safe_auto cleanups:
- Drop the redundant `as T` cast at the end of pickAliasedArg; both
  ternary branches were already `T | undefined`, matching the declared
  return type.
- Replace the unchecked `JSON.parse(...) as Record<string, unknown>` in
  loadSettingsConfig with `parsed: unknown` + an isRecord guard. Same
  runtime behavior, honest type at the boundary.

Public API unchanged. 115/115 tests pass, npm run check clean.
